show the instruction UI only once
persist between scene but not refresh
TODO: maybe add a keybind to show this UI
