+++ Assets/Script/PlayerStats.cs
@@ -1,0 +1,7 @@
+namespace PlayerStats
+{
+    public static class UIStats
+    {
+        public static bool playerSeenInstruction { get; set; } = false;
+    }
+}
