+++ Assets/Script/PlayerStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 25151c126d9c5d04ab3555c1cea7f7b5
+guid: ae66e74969bcc76eabbf399aa1d096b2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
