--- Assets/Script/closeInstructionUI.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class closeInstructionUI : MonoBehaviour
-{
-    public GameObject InstructionUI;
-    // Start is called before the first frame update
-    void Start()
-    {
-        InstructionUI.SetActive(false);
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
