+++ Assets/Script/startGame.cs
@@ -1,10 +1,12 @@
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using PlayerStats;
 
-public class loadLevel : MonoBehaviour
+public class startGame : MonoBehaviour
 {
-    public void LoadLevel1()
+    public void StartGame()
     {
+        UIStats.playerSeenInstruction = false;
         // 使用 SceneManager.LoadScene 來讀取 "Level1" 場景
         SceneManager.LoadScene("Level1");
     }
