+++ Assets/Script/startGame.cs.meta
